Change the way to check if a term can change its name
In the previous version of the code, you could only change the name of a term if it was a variable.

However, if we extend the code (for instance, to include labelled nulls), you are forcing the labelled nulls to override the isVariable() function to return true to change their name!

In the new version, you have a new function called nameIsModifiable that can be override by the functions freely.